requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1
Newtonsoft.Json==13.0.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
asyncpg==0.29.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
ta==0.11.0
yfinance==0.2.32
newsapi==0.1.2
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
pyjwt==2.8.1
cryptography==41.0.7
cors==1.0.1
